+++ extensions/menubar-weather/src/types/preferences.ts
@@ -6,6 +6,10 @@ export interface Preferences {
   windSpeedUnits: string;
   precipitationUnits: string;
   tempType: string;
+  menuUVI: boolean;
+  menuPressure: boolean;
+  menuHumidity: boolean;
+  menuWind: boolean;
   showSun: boolean;
   showUVI: boolean;
   showLocation: boolean;